Update v1_11_1
Added frame sizes to be assigned to each instrument type in the layouts menu.

Added a custom .lrw file that can be downloaded and reuploaded. This saves just like a save state does on the local computer. Saving any data that you have put into the worknotes or layout menu's.